# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Core App Rest",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/roles": {
            "get": {
                "tags": [
                    "Admin Role"
                ],
                "summary": "get all roles",
                "description": "get all list roles",
                "operationId": "ce386107ed2fe1c04b9ad915d649432c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/direction"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Specify sorting order.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "created_at"
                            ]
                        },
                        "example": "id"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Apply filter by name role.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Apply filter by code role.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SUPER_ADMIN",
                                "USER"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return All Roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "properties": {
                                                "pagination": {
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Role"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Role"
                ],
                "summary": "create role",
                "description": "create new role",
                "operationId": "c3eb2a775782636946562aa2f6ab60f4",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/x-locale"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "super role"
                                    },
                                    "standard": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Created Role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/admin/roles/{role}": {
            "get": {
                "tags": [
                    "Admin Role"
                ],
                "summary": "Get one role",
                "description": "Fetches a single role by ID",
                "operationId": "c95013f309e5a07c29d7975083fbe01b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/x-locale"
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "the role to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Get One Role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Role"
                ],
                "summary": "update role",
                "description": "update one role",
                "operationId": "e9b409c3d25bf674456a59fd72e9530e",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/x-locale"
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "role to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "super role"
                                    },
                                    "standard": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Updated Role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Role"
                ],
                "summary": "delete role",
                "description": "delete a single role by ID",
                "operationId": "22d1bb549b14281650be4ca5211e342f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/x-locale"
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "role to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Deleted Role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pagination": {
                "description": "Pagination schema",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "pagination": {
                        "properties": {
                            "page": {
                                "type": "integer",
                                "example": 1
                            },
                            "pages": {
                                "type": "integer",
                                "example": 10
                            },
                            "totalItems": {
                                "type": "integer",
                                "example": 200
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Role": {
                "description": "Role schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "role-name"
                    },
                    "standard": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "direction": {
                "name": "direction",
                "in": "query",
                "description": "Specify sorting order.",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                },
                "example": "asc"
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "Determine page number.",
                "required": false,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            },
            "limit": {
                "name": "limit",
                "in": "query",
                "description": "Determine number of items per page.",
                "required": false,
                "schema": {
                    "type": "integer",
                    "enum": [
                        10,
                        20,
                        30,
                        40
                    ]
                },
                "example": 10
            },
            "x-locale": {
                "name": "x-locale",
                "in": "header",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "en",
                        "ar"
                    ]
                },
                "example": "en"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}